requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
PyJWT==2.3.0
pymongo==3.12.1
pytz==2022.1
sqlparse==0.2.4
tzdata==2022.1
